Move to a dedicated settings plugin file
